tied up all loose ends
