+++ .dockerignore
@@ -5,3 +5,4 @@ node_modules
 Dockerfile
 LICENSE
 /deploy
+_*
