--- src/server/ui/index.ts
@@ -1,3 +1,0 @@
-export * from './Analytics';
-export * from './InmateCard';
-export * from './JsonBox';
